requirements.txt
Django==5.0.6
djangorestframework>=3.15.2
mysqlclient==2.2.4
python-dotenv==1.0.1
djangorestframework-simplejwt==5.3.1
pytz==2024.1
mysqlclient-2.2.8==2.2.8